Update with neo repos
Picked some of the various NEO repos with reasonably active development. Not sure exactly what criteria you guys wish to use to decide which repos to include, so here's a bunch to get started that you can do whatever with. Can suggest some more depending on what you are looking for.